requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.18
numpy==1.24.3
pillow==10.0.0
streamlit==1.28.0
streamlit-webrtc==0.47.1
av==10.0.0